script for MEMA and bruxism fully operational
